DMP-3872 - change point where darNotify is async
